Create a main component for users to learn about Startup Force and sign up or log out.
