+++ server/index.js
@@ -27,8 +27,8 @@ app.use('/', scheduleRouter);
 app.use('/', reportsRouter);
 app.use('/', loginRouter);
 
-
-app.use('/*', (req, res) => res.sendFile(path.join(__dirname, '/../client/dist')));
+app.use('/*', (req, res) => res.redirect('/'));
+//app.use('/*', (req, res) => res.sendFile(path.join(__dirname, '/../client/dist')));
 // initialize Passport
 //app.use(session({ secret: '369lex' }));
 app.use(passport.initialize());
